fix single element sampler array reflection - make lights demo run on pi

+++ system_modules/naprender/src/descriptorsetallocator.cpp
@@ -27,7 +27,7 @@ namespace nap
 
 	VkDescriptorSet DescriptorSetAllocator::allocate(VkDescriptorSetLayout layout, int numUBODescriptors, int numSSBODescriptors, int numSamplerDescriptors)
 	{
-		uint64_t key = ((uint64_t)numUBODescriptors) << 32 | ((uint64_t)numSSBODescriptors << 16) | numSamplerDescriptors;
+		uint64_t key = ((uint64_t)numUBODescriptors) << 32 | ((uint64_t)numSSBODescriptors << 16) | ((uint64_t)numSamplerDescriptors);
 
 		DescriptorPool* free_descriptor_pool = nullptr;
 		DescriptorPoolMap::iterator pos = mDescriptorPools.find(key);
